+++ cgi-bin/index.py
@@ -64,9 +64,16 @@ print(f"""
     <link rel="stylesheet" href="../styles/styles.css">
     <link rel="stylesheet" href="../styles/styles2.css">
     <link rel="stylesheet" href="../styles/modals.css">
+    <link rel="stylesheet" href="https://unpkg.com/leaflet@1.7.1/dist/leaflet.css"
+        integrity="sha512-xodZBNTC5n17Xt2atTPuE1HxjVMSvLVW9ocqUKLsCC5CXdbqCmblAshOMAS6/keqq/sMZMZ19scR4PsZChSR7A=="
+        crossorigin=""/>
     <script src="https://code.jquery.com/jquery-3.6.0.min.js"
             integrity="sha256-/xUj+3OJU5yExlq6GSYGSHk7tPXikynS7ogEvDej/m4="
             crossorigin="anonymous"></script>
+    <script src="https://unpkg.com/leaflet@1.7.1/dist/leaflet.js"
+        integrity="sha512-XQoYMqMTK8LvdxXYG3nZ448hOEQiglfqkJs1NOQV44cWnUrBc8PkAOcXy20w0vlaXaVUearIOBhiXZ5V3ynxwA=="
+        crossorigin=""></script>
+    
 </head>
 <body>
 
@@ -124,6 +131,7 @@ print(f"""
 
     </table>
 </div>
+<div id="map_id"></div>
 <a href="../templates/estadísticas.html">
     <div class="box round_square">
         <div class="center bold">Ver estadísticas.</div>
@@ -136,6 +144,7 @@ print(f"""
 
 <script src="../scripts/functions.js"></script>
 <script src="../scripts/validations.js"></script>
+<script src="../scripts/map.js"></script>
 
 </body>
 </html>